2025/aidesign/pyproject.toml
[project]
name = "aidesign"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]

[tool.uv]
environments = ["sys_platform == 'darwin'"]